// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Provider = {
  STRIPE: 'STRIPE',
  SSLCOMMERZ: 'SSLCOMMERZ'
} as const

export type Provider = (typeof Provider)[keyof typeof Provider]


export const PaymentStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const Availability = {
  AVAILABLE: 'AVAILABLE',
  RENTED: 'RENTED',
  UNAVAILABLE: 'UNAVAILABLE'
} as const

export type Availability = (typeof Availability)[keyof typeof Availability]


export const RequestStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  CANCELLED: 'CANCELLED'
} as const

export type RequestStatus = (typeof RequestStatus)[keyof typeof RequestStatus]


export const Roles = {
  TENANT: 'TENANT',
  LANDLORD: 'LANDLORD',
  ADMIN: 'ADMIN'
} as const

export type Roles = (typeof Roles)[keyof typeof Roles]


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  BLOCKED: 'BLOCKED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]